+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added cert options to the GraphAppToolkit send email.
 - Updated auth methods to invoke needed permissions only.
+- Added private functions to handle existing certs and secrets.
 
 ### Fixed